更新后台代码

+++ src/main/java/com/project/gelingeducation/common/annotation/Cache.java
@@ -6,5 +6,5 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Cache {
     //过期时间
-    long time() default 30*60;
+    int time() default 30*60;
 }
